Rename model->models for consistency
